requirements.txt
# Core dependencies
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.1
pydantic-settings==2.1.0
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
chromadb==0.4.22

# AI/ML dependencies
openai==1.12.0
google-generativeai==0.3.2
python-dotenv==1.0.1

# Testing
pytest==8.0.1
pytest-asyncio==0.23.5
httpx==0.26.0

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Additional dependencies
email-validator==2.1.1 